pyproject.toml
[project]
name = "wonder-scraper"
version = "0.1.0"
description = "eBay scraper for TCG Wonders of the First"
authors = [
    {name = "Cody",email = "cody@example.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "sqlmodel (>=0.0.27,<0.0.28)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pdfplumber (>=0.11.8,<0.12.0)",
    "pydoll-python @ git+https://github.com/autoscrape-labs/pydoll.git",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "apscheduler (>=3.11.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
